0024234: occt master is not compiled by VC++ 2005 (vc8 32/64 bit TKBO)
Fixed building by VS 2005 and VS 2013 (Express):
- macro _WIN32_WINNT defined in BOPCol_TBB.hxx for building with TBB on VS 2005
- compiler warnings specific for VS 2005 (vc8) fixed (explicit casts added in assignment expressions involving small integer types)
- definition of functions atanh, acosh, asinh is avoided for vc12 and later (they are provided by CRT in vc12)
- math constants (M_PI and similar) are defined by including math.h rather than by custom code
- misuse of stream operator << fixed in a few places

+++ src/XmlObjMgt/XmlObjMgt.cxx
@@ -92,7 +92,7 @@ void SprintfExtStr(char * out, const TCollection_ExtendedString& theString) {
   while (len) {
     for(int j = 0,k=3; j<4; j++,k--) {
       unsigned short v = *(p+i) & mask[j];//x000
-      v = v >> (4*k);
+      v = (unsigned short)(v >> (4*k));
       if(v < 10)
 	v |= 0x30;
       else
